Modify : member-회원 탈퇴 기능에 비밀번호 검증 구현 및 아이디 찾기 반환 글 일부 수정

+++ src/main/java/com/prgrms/ijuju/domain/member/service/MemberService.java
@@ -160,10 +160,15 @@ public class MemberService {
 
     // 회원 탈퇴
     @Transactional
-    public void delete(Long id) {
+    public void delete(Long id, String pw) {
         Optional<Member> opMember = memberRepository.findById(id);
         if (opMember.isPresent()) {
             Member member = opMember.get();
+
+            if (!passwordEncoder.matches(pw, member.getPw())) {
+                throw new SecurityException("비밀번호가 일치하지 않습니다.");
+            }
+
             memberRepository.delete(member);
         } else {
             throw MemberException.MEMBER_NOT_REMOVED.getMemberTaskException();
